add rating of media and small changes

+++ backend/social/review/models.go
@@ -1,18 +1,18 @@
 package review
 
 type ReviewRequest struct {
-	MediaID uint   `json:"media_id"`
-	Rating  uint8  `json:"rating"`
-	Body    string `json:"body"`
-	Rewatch bool   `json:"rewatch"`
-	Date    string `json:"date"`
+	MediaID uint    `json:"media_id"`
+	Rating  float32 `json:"rating"`
+	Body    string  `json:"body"`
+	Rewatch bool    `json:"rewatch"`
+	Date    string  `json:"date"`
 }
 
 type ReviewResponse struct {
-	ID      uint   `json:"id"`
-	MediaID uint   `json:"media_id"`
-	Rating  uint8  `json:"rating"`
-	Body    string `json:"body"`
-	Rewatch bool   `json:"rewatch"`
-	Date    string `json:"date"`
+	ID      uint    `json:"id"`
+	MediaID uint    `json:"media_id"`
+	Rating  float32 `json:"rating"`
+	Body    string  `json:"body"`
+	Rewatch bool    `json:"rewatch"`
+	Date    string  `json:"date"`
 }
